+++ src/main/java/engine/ControlledComp.java
@@ -1,8 +1,5 @@
 package engine;
 
-import engine.Component;
-import engine.network.CharacterInputData;
-
 /**
  * Created by eirik on 05.07.2017.
  */
